api/requirements.txt
SQLAlchemy~=2.0.21
fastapi~=0.103.2
python-jose~=3.3.0
elasticsearch~=8.10.0
pydantic~=2.4.2
passlib~=1.7.4
alembic~=1.12.0
uvicorn[standard]==0.23.2
psycopg2-binary==2.9.9
python-multipart==0.0.6
